pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obsidian-markdown-to-pdf"
version = "1.0.0"
description = "Convert Obsidian-flavored markdown files to PDF"
requires-python = ">=3.12,<3.13"
dependencies = [
    "markdown>=3.5",
    "obsidian-media>=1.0.0",
    "weasyprint>=62.0",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "types-Markdown",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["main.py", "src"]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "types-markdown>=3.10.2.20260211",
]
